pyproject.toml
[project]
name = "colonoscopy-triage-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.2",
    "langchain>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
]
